+++ stroom-config/stroom-config-global-impl/src/main/java/stroom/config/global/impl/ConfigProvidersModule.java
@@ -594,15 +594,6 @@ public class ConfigProvidersModule extends AbstractModule {
     @Generated("stroom.config.global.impl.GenerateConfigProvidersModule")
     @Provides
     @SuppressWarnings("unused")
-    stroom.searchable.impl.SearchableConfig getSearchableConfig(
-            final ConfigMapper configMapper) {
-        return configMapper.getConfigObject(
-                stroom.searchable.impl.SearchableConfig.class);
-    }
-
-    @Generated("stroom.config.global.impl.GenerateConfigProvidersModule")
-    @Provides
-    @SuppressWarnings("unused")
     stroom.security.identity.config.EmailConfig getEmailConfig(
             final ConfigMapper configMapper) {
         return configMapper.getConfigObject(