+++ stroom-dashboard/stroom-dashboard-impl/src/main/java/stroom/dashboard/impl/DashboardConfig.java
@@ -18,7 +18,7 @@ public class DashboardConfig extends AbstractConfig implements IsStroomConfig {
 
     public DashboardConfig() {
         activeQueriesCache = CacheConfig.builder()
-                .maximumSize(100L)
+                .maximumSize(1000L)
                 .expireAfterAccess(StroomDuration.ofMinutes(1))
                 .build();
     }