+++ stroom-dashboard/stroom-dashboard-impl/src/main/java/stroom/dashboard/impl/SearchRequestMapper.java
@@ -24,7 +24,6 @@ import stroom.dashboard.impl.vis.VisSettings.Structure;
 import stroom.dashboard.impl.vis.VisSettings.Tab;
 import stroom.dashboard.impl.visualisation.VisualisationStore;
 import stroom.dashboard.shared.ComponentResultRequest;
-import stroom.dashboard.shared.DashboardQueryKey;
 import stroom.dashboard.shared.DashboardSearchRequest;
 import stroom.dashboard.shared.TableResultRequest;
 import stroom.dashboard.shared.VisComponentSettings;
@@ -34,7 +33,6 @@ import stroom.query.api.v2.ExpressionOperator;
 import stroom.query.api.v2.Format;
 import stroom.query.api.v2.Param;
 import stroom.query.api.v2.Query;
-import stroom.query.api.v2.QueryKey;
 import stroom.query.api.v2.ResultRequest;
 import stroom.query.api.v2.ResultRequest.ResultStyle;
 import stroom.query.api.v2.SearchRequest;
@@ -69,14 +67,13 @@ public class SearchRequestMapper {
         this.visualisationStore = visualisationStore;
     }
 
-    public SearchRequest mapRequest(final DashboardQueryKey queryKey,
-                                    final DashboardSearchRequest searchRequest) {
+    public SearchRequest mapRequest(final DashboardSearchRequest searchRequest) {
         if (searchRequest == null) {
             return null;
         }
 
         return SearchRequest.builder()
-                .key(new QueryKey(queryKey.toUuid()))
+                .key(searchRequest.getQueryKey())
                 .query(mapQuery(searchRequest))
                 .resultRequests(mapResultRequests(searchRequest))
                 .dateTimeSettings(searchRequest.getDateTimeSettings())