+++ stroom-query/stroom-query-common/src/main/java/stroom/query/common/v2/ResultStoreConfig.java
@@ -1,8 +1,10 @@
 package stroom.query.common.v2;
 
+import stroom.util.cache.CacheConfig;
 import stroom.util.io.ByteSize;
 import stroom.util.shared.AbstractConfig;
 import stroom.util.shared.IsStroomConfig;
+import stroom.util.time.StroomDuration;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -25,6 +27,9 @@ public class ResultStoreConfig extends AbstractConfig implements IsStroomConfig 
 
     private final ResultStoreLmdbConfig lmdbConfig;
 
+    private final CacheConfig searchResultCache;
+    private final String storeSize;
+
     public ResultStoreConfig() {
         maxPutsBeforeCommit = 100_000;
         offHeapResults = true;
@@ -36,6 +41,12 @@ public class ResultStoreConfig extends AbstractConfig implements IsStroomConfig 
         maxPayloadSize = ByteSize.ofGibibytes(1);
 
         lmdbConfig = new ResultStoreLmdbConfig();
+
+        searchResultCache = CacheConfig.builder()
+                .maximumSize(10000L)
+                .expireAfterAccess(StroomDuration.ofMinutes(10))
+                .build();
+        storeSize = "1000000,100,10,1";
     }
 
     @SuppressWarnings("unused")
@@ -47,7 +58,9 @@ public class ResultStoreConfig extends AbstractConfig implements IsStroomConfig 
                              @JsonProperty("minPayloadSize") final ByteSize minPayloadSize,
                              @JsonProperty("maxPayloadSize") final ByteSize maxPayloadSize,
                              @JsonProperty("valueQueueSize") final int valueQueueSize,
-                             @JsonProperty("lmdb") final ResultStoreLmdbConfig lmdbConfig) {
+                             @JsonProperty("lmdb") final ResultStoreLmdbConfig lmdbConfig,
+                             @JsonProperty("searchResultCache") final CacheConfig searchResultCache,
+                             @JsonProperty("storeSize") final String storeSize) {
         this.maxPutsBeforeCommit = maxPutsBeforeCommit;
         this.offHeapResults = offHeapResults;
         this.minValueSize = minValueSize;
@@ -56,6 +69,8 @@ public class ResultStoreConfig extends AbstractConfig implements IsStroomConfig 
         this.maxPayloadSize = maxPayloadSize;
         this.valueQueueSize = valueQueueSize;
         this.lmdbConfig = lmdbConfig;
+        this.searchResultCache = searchResultCache;
+        this.storeSize = storeSize;
     }
 
     @Min(0)
@@ -105,16 +120,29 @@ public class ResultStoreConfig extends AbstractConfig implements IsStroomConfig 
         return lmdbConfig;
     }
 
+    @JsonProperty("searchResultCache")
+    public CacheConfig getSearchResultCache() {
+        return searchResultCache;
+    }
+
+    @JsonPropertyDescription("The maximum number of search results to keep in memory at each level.")
+    public String getStoreSize() {
+        return storeSize;
+    }
+
     @Override
     public String toString() {
         return "ResultStoreConfig{" +
                 "maxPutsBeforeCommit=" + maxPutsBeforeCommit +
                 ", offHeapResults=" + offHeapResults +
+                ", valueQueueSize=" + valueQueueSize +
                 ", minValueSize=" + minValueSize +
                 ", maxValueSize=" + maxValueSize +
                 ", minPayloadSize=" + minPayloadSize +
                 ", maxPayloadSize=" + maxPayloadSize +
                 ", lmdbConfig=" + lmdbConfig +
+                ", searchResultCache=" + searchResultCache +
+                ", storeSize='" + storeSize + '\'' +
                 '}';
     }
 }