+++ stroom-query/stroom-query-common/src/main/java/stroom/query/common/v2/SearchResponseCreatorFactory.java
@@ -2,7 +2,7 @@ package stroom.query.common.v2;
 
 import javax.inject.Inject;
 
-public class SearchResponseCreatorFactory {
+public final class SearchResponseCreatorFactory {
 
     private final SizesProvider sizesProvider;
 
@@ -14,7 +14,7 @@ public class SearchResponseCreatorFactory {
     /**
      * @param store The underlying store to use for creating the search responses.
      */
-    public SearchResponseCreator create(final Store store) {
-        return new SearchResponseCreator(sizesProvider, store);
+    public SearchResponseCreator create(final String userId, final Store store) {
+        return new SearchResponseCreator(userId, sizesProvider, store);
     }
 }