+++ stroom-statistics/stroom-statistics-impl-sql/src/main/java/stroom/statistics/impl/sql/search/SQLStatisticSearchModule.java
@@ -17,18 +17,11 @@
 package stroom.statistics.impl.sql.search;
 
 import stroom.datasource.api.v2.DataSourceProvider;
-import stroom.job.api.ScheduledJobsBinder;
 import stroom.statistics.impl.sql.StatisticsQueryService;
-import stroom.util.RunnableWrapper;
 import stroom.util.guice.GuiceUtil;
 import stroom.util.guice.RestResourcesBinder;
-import stroom.util.shared.Clearable;
 
 import com.google.inject.AbstractModule;
-
-import javax.inject.Inject;
-
-import static stroom.job.api.Schedule.ScheduleType.PERIODIC;
 
 public class SQLStatisticSearchModule extends AbstractModule {
 
@@ -37,28 +30,10 @@ public class SQLStatisticSearchModule extends AbstractModule {
         bind(StatisticsQueryService.class).to(StatisticsQueryServiceImpl.class);
         bind(StatisticsSearchService.class).to(StatisticsSearchServiceImpl.class);
 
-        GuiceUtil.buildMultiBinder(binder(), Clearable.class)
-                .addBinding(SqlStatisticsSearchResponseCreatorManager.class);
-
         GuiceUtil.buildMultiBinder(binder(), DataSourceProvider.class)
                 .addBinding(StatisticsQueryServiceImpl.class);
 
         RestResourcesBinder.create(binder())
                 .bind(SqlStatisticsQueryResourceImpl.class);
-
-        ScheduledJobsBinder.create(binder())
-                .bindJobTo(EvictExpiredElements.class, builder -> builder
-                        .name("Evict expired elements")
-                        .schedule(PERIODIC, "10s")
-                        .managed(false));
-    }
-
-    private static class EvictExpiredElements extends RunnableWrapper {
-
-        @Inject
-        EvictExpiredElements(
-                final SqlStatisticsSearchResponseCreatorManager sqlStatisticsSearchResponseCreatorManager) {
-            super(sqlStatisticsSearchResponseCreatorManager::evictExpiredElements);
-        }
     }
 }